+++ packages/reactive_file_picker/lib/reactive_file_picker.dart
@@ -112,6 +112,14 @@ class ReactiveFilePicker<T>
     bool lockParentWindow = false,
     double disabledOpacity = 0.5,
     String? initialDirectory,
+    Widget Function(BuildContext context, String error)? errorBuilder,
+
+    // input decorator props
+    TextStyle? baseStyle,
+    TextAlign? textAlign,
+    TextAlignVertical? textAlignVertical,
+    bool expands = false,
+    MouseCursor cursor = SystemMouseCursors.basic,
   }) : super(
           builder: (field) {
             final value = field.value ?? MultiFile<T>();
@@ -157,22 +165,87 @@ class ReactiveFilePicker<T>
               }
             }
 
-            return InputDecorator(
-              decoration: effectiveDecoration.copyWith(
-                errorText: field.errorText ?? pickerError,
-                enabled: field.control.enabled,
-              ),
-              child: IgnorePointer(
-                ignoring: !field.control.enabled,
-                child: Opacity(
-                  opacity: field.control.enabled ? 1 : disabledOpacity,
-                  child: filePickerBuilder?.call(pickFile, value, (files) {
-                    field.control.markAsTouched();
-                    field.didChange(files);
-                  }),
-                ),
+            final errorText = field.errorText;
+
+            final isEmptyValue =
+                field.value == null || (field.value?.platformFiles.isEmpty == true && field.value?.files.isEmpty == true);
+
+            return IgnorePointer(
+              ignoring: !field.control.enabled,
+              child: MouseRegion(
+                cursor: cursor,
+                child: HoverBuilder(builder: (context, isHovered) {
+                  return InputDecorator(
+                    isHovering: isHovered,
+                    baseStyle: baseStyle,
+                    textAlign: textAlign,
+                    textAlignVertical: textAlignVertical,
+                    expands: expands,
+                    isEmpty: isEmptyValue,
+                    decoration: effectiveDecoration.copyWith(
+                      enabled: field.control.enabled,
+                      errorText: errorBuilder == null
+                          ? field.errorText ?? pickerError
+                          : null,
+                      error: errorBuilder != null && errorText != null
+                          ? DefaultTextStyle.merge(
+                              style: Theme.of(field.context)
+                                  .textTheme
+                                  .bodySmall
+                                  ?.copyWith(
+                                    color:
+                                        Theme.of(field.context).colorScheme.error,
+                                  )
+                                  .merge(effectiveDecoration.errorStyle),
+                              child: errorBuilder.call(
+                                field.context,
+                                errorText,
+                              ),
+                            )
+                          : null,
+                    ),
+                    child: Opacity(
+                      opacity: field.control.enabled ? 1 : disabledOpacity,
+                      child: filePickerBuilder?.call(pickFile, value, (files) {
+                        field.control.markAsTouched();
+                        field.didChange(files);
+                      }),
+                    ),
+                  );
+                }),
               ),
             );
           },
         );
+}
+
+class HoverBuilder extends StatefulWidget {
+  const HoverBuilder({
+    required this.builder,
+    super.key,
+  });
+
+  final Widget Function(BuildContext context, bool isHovered) builder;
+
+  @override
+  _HoverBuilderState createState() => _HoverBuilderState();
+}
+
+class _HoverBuilderState extends State<HoverBuilder> {
+  bool _isHovered = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return MouseRegion(
+      onEnter: (event) => _onHoverChanged(enabled: true),
+      onExit: (event) => _onHoverChanged(enabled: false),
+      child: widget.builder(context, _isHovered),
+    );
+  }
+
+  void _onHoverChanged({required bool enabled}) {
+    setState(() {
+      _isHovered = enabled;
+    });
+  }
 }